Made genversion.py use the version of Python found by CMake.
In non-standard builds a "python2" executable may not be just available to
the genversion.py script out of the box. Python may be sitting in some
non-standard place. To make the genversion.py script easy to use in such
cases as well, it is now configured during the CMake configuration to use
the version of Python found by CMake's FindPythonInterp.cmake module.

For this to work I had to also modify genversion.py a little bit, so that
configure_file(...) would not break the script.