Made all Python scripts use the Python interpreter found by CMake.
To be consistent with the update made to genversion.py, updated the code
to create all the other python scripts of the project such that they would
also always use the same python version that CMake found.

For this to work I also updated the .gitignore file, since the scripts that
were previously part of the repository, are now created during the CMake
configuration.